Removed unused methods
